+++ Source/Public/Add-MyRMClient.ps1
@@ -38,7 +38,7 @@ function Add-MyRMClient {
     PS> Add-MyRMClient -Name SSH -Executable "ssh.exe" -Arguments "-l <user> -p <port> <host>" -DefaultPort 22
 
     .EXAMPLE
-    PS> Add-MyRMClient -Name MyCustomClient -Executable "client.exe" -Arguments "--hostname <host> --port <port>" -DefaultPort 666 -Description "My custom client"
+    PS> Add-MyRMClient -Name MyCustomClient -Executable "client.exe" -Arguments "--hostname <host> --port <port>" -DefaultPort 666 -DefaultScope External -Description "My custom client"
 
     #>
 
@@ -74,7 +74,6 @@ function Add-MyRMClient {
         [UInt16] $DefaultPort,
 
         [Parameter(
-            Mandatory = $false,
             HelpMessage = "Default scope in which a connection will be invoked."
         )]
         [ValidateNotNullOrEmpty()]
@@ -100,6 +99,7 @@ function Add-MyRMClient {
             $DefaultScope,
             $Description
         )
+
         if (
             $PSCmdlet.ShouldProcess(
                 "Inventory file {0}" -f $Inventory.Path,