+++ Source/Public/Invoke-MyRMConnection.ps1
@@ -11,6 +11,12 @@ function Invoke-MyRMConnection {
     .PARAMETER Name
     Name of the connection.
 
+    .PARAMETER Client
+    Name of the client to use to initiate the connection.
+
+    .PARAMETER User
+    Name of the user to connect with.
+
     .PARAMETER Scope
     Scope in which the connection will be invoked.
 
@@ -24,7 +30,7 @@ function Invoke-MyRMConnection {
     PS> Invoke-MyRMConnection myconn
 
     .EXAMPLE
-    PS> Invoke-MyRMConnection -Name myconn
+    PS> Invoke-MyRMConnection -Name myconn -Client SSH -User root -Scope Console
 
     #>
 
@@ -36,13 +42,26 @@ function Invoke-MyRMConnection {
             Mandatory = $true,
             HelpMessage = "Name of the connection."
         )]
-        [ValidateNotNullOrEmpty()]
-        [ValidateSet( [ValidateConnectionName] )]
+        [ValidateSet([ValidateConnectionName])]
         [string] $Name,
+
+        [Parameter(
+            HelpMessage = "Name of the client to use to initiate the connection."
+        )]
+        [ValidateSet([ValidateClientName])]
+        [Alias("c")]
+        [string] $Client,
+
+        [Parameter(
+            HelpMessage = "Name of the user to connect with."
+        )]
+        [Alias("u")]
+        [string] $User,
 
         [Parameter(
             HelpMessage = "Scope in which the connection will be invoked."
         )]
+        [Alias("x")]
         [Scopes] $Scope
     )
 
@@ -52,27 +71,71 @@ function Invoke-MyRMConnection {
     }
 
     process {
-        $Connection = $Inventory.Connections | Where-Object -Property Name -EQ $Name
-        if ($PSCmdlet.ShouldProcess($Connection.ToString(), "Initiate connection")) {
-            if ($Connection.Client.RequiresUser) {
-                $User = Read-Host -Prompt ("Username" -f $Connection.Hostname)
-                $Arguments = $Connection.GenerateArgs($User)
+        $Invocation = @{}
+
+        $Invocation.Connection = $Inventory.GetConnection($Name)
+        Write-Debug -Message ("Invoke connection {0}" -f $Invocation.Connection.ToString())
+
+        $Invocation.Client = if ($Client) {
+            $Inventory.GetClient($Client)
+        }
+        else {
+            $Inventory.GetClient($Invocation.Connection.DefaultClient)
+        }
+        Write-Debug -Message ("Invoke connection with client {0}" -f $Invocation.Client.ToString())
+
+        $Invocation.Port = if ($Invocation.Connection.IsDefaultPort()) {
+            $Invocation.Client.DefaultPort
+        }
+        else {
+            $Invocation.Connection.Port
+        }
+        Write-Debug -Message ("Invoke connection on port {0}" -f $Invocation.Port)
+
+        $Invocation.Executable = $Invocation.Client.Executable
+        $Invocation.Arguments = if ($Invocation.Client.RequiresUser) {
+            if ($User) {
+                $Invocation.Client.GenerateArgs(
+                    $Invocation.Connection.Hostname,
+                    $Invocation.Port,
+                    $User
+                )
+            }
+            elseif ($Invocation.Connection.DefaultUser) {
+                $Invocation.Client.GenerateArgs(
+                    $Invocation.Connection.Hostname,
+                    $Invocation.Port,
+                    $Invocation.Connection.DefaultUser
+                )
             }
             else {
-                $Arguments = $Connection.GenerateArgs()
+                throw "Cannot invoke connection: A user must be specified."
             }
+        }
+        else {
+            $Invocation.Client.GenerateArgs(
+                $Invocation.Connection.Hostname,
+                $Invocation.Port
+            )
+        }
+        $Invocation.Command = "{0} {1}" -f $Invocation.Executable, $Invocation.Arguments
+        Write-Debug -Message ("Invoke connection with command `"{0}`"" -f $Invocation.Command)
 
-            $Process = @{
-                FilePath     = $Connection.Client.Executable
-                ArgumentList = $Arguments
-            }
-            switch ($Scope) {
+        $Invocation.Scope = if ($Scope) {
+            $Scope
+        }
+        else {
+            $Invocation.Client.DefaultScope
+        }
+        Write-Debug -Message ("Invoke connection in scope `"{0}`"" -f $Invocation.Scope)
+
+        if ($PSCmdlet.ShouldProcess($Invocation.Connection.ToString(), "Initiate connection")) {
+            switch ($Invocation.Scope) {
                 ([Scopes]::Console) {
-                    $Command = "{0} {1}" -f $Process.FilePath, $Process.ArgumentList
-                    Invoke-Expression -Command $Command
+                    Invoke-Expression -Command $Invocation.Command
                 }
                 ([Scopes]::External) {
-                    Start-Process @Process
+                    Start-Process -FilePath $Invocation.Executable -ArgumentList $Invocation.Arguments
                 }
                 ([Scopes]::Undefined) {
                     throw "Cannot invoke connection: Scope is undefined."